adding genres component
